Refactoring window events/size methods to Window.c
Also adding missing sources to src/Makefile.am.
Fixing small bug in Input.c
