+++ src/Internal.h
@@ -43,6 +43,8 @@ bool EventQueue_IsEmpty();
 void Input_StartAsyncThread();
 void Input_StopAsyncThread();
 
+void Window_StartSizeListener();
+
 void Log_Open();
 void Log_Info(const char *message, ...);
 void Log_Error(const char *message, ...);
