+++ src/Makefile.am
@@ -5,5 +5,8 @@ libbulwark_la_SOURCES = \
   Bulwark.c \
   Event.c \
   EventQueue.c \
+  Window.c \
+  Input.c \
   Internal.h \
+  Log.c \
   Sound.c
